+++ comicbox/box/archive/init.py
@@ -3,10 +3,6 @@
 import re
 from functools import wraps
 from logging import getLogger
-from sys import maxsize
-
-from py7zr import SevenZipFile
-from py7zr.io import BytesIOFactory
 
 from comicbox.box.init import ComicboxInit
 from comicbox.box.types import ArchiveType
@@ -58,6 +54,4 @@ class ComicboxArchiveInit(ComicboxInit):
         if not self._archive:
             reason = f"Unable to make archive from class {self._archive_cls}"
             raise ValueError(reason)
-        if self._archive_cls == SevenZipFile:
-            self._7zfactory: BytesIOFactory | None = BytesIOFactory(maxsize)
         return self._archive